+++ lib/one_logger.dart
@@ -5,6 +5,7 @@ library simple_logger;
 
 export 'src/logger.dart';
 export 'src/helper.dart' hide defaultDateFormat;
+export 'src/options/filter.dart';
 export 'src/options/options.dart';
 export 'src/options/style.dart';
 export 'src/loki/options.dart';