+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.0.8
+
+- Added `bool enableAnsi` to `LoggerOptions`.
+
+## 1.0.7
+
+- Fixed missing `Filter` export.
+
+## 1.0.6
+
+- Dependency updates.
+
+## 1.0.5
+
+- Minor improvements.
+
 ## 1.0.4
 
 - Introduced `LoggerFilter`. Now logs can be filtered by `Level`